condition state utils renamed
